requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-groq
chromadb
sentence-transformers
transformers
pypdf
python-dotenv
langchain-text-splitters